Refactoring(services): changing success scenarios
