+++ backend/features/cadastroServico.feature
@@ -3,27 +3,27 @@ Feature: Serviço de Cadastro de Usuários
   Scenario: Sucesso no cadastro de usuário
     Given o usuário deseja se cadastrar
     When ele informa o nome "Demosténes"
-    And ele informa o CPF "126.456.789-00"
-    And ele informa o email "demostenes@example.com"
+    And ele informa o CPF "126.455.789-00"
+    And ele informa o email "demostenessouza@example.com"
     And ele informa se é professor "N"
     And ele informa a senha "SecurePassword123"
     And ele informa a confirmação da senha "SecurePassword123"
     And ele envia uma requisição POST para "/api/cadastro"
     Then a resposta deve conter a mensagem "Cadastro criado com sucesso!"
-    And o status code deve ser 201
+    And o status code deve ser "201"
 
   Scenario: Sucesso no cadastro de professor
     Given o usuário deseja se cadastrar
-    When ele informa o nome "Paula"
-    And ele informa o CPF "321.879.789-33"
-    And ele informa o email "vanessa@example.com"
+    When ele informa o nome "Tatiana"
+    And ele informa o CPF "331.879.789-33"
+    And ele informa o email "vanessasilva@example.com"
     And ele informa se é professor "S"
-    And ele informa o SIAPE "101010"
+    And ele informa o SIAPE "101110"
     And ele informa a senha "12345678"
     And ele informa a confirmação da senha "12345678"
     And ele envia uma requisição POST para "/api/cadastro"
     Then a resposta deve conter a mensagem "Cadastro criado com sucesso!"
-    And o status code deve ser 201
+    And o status code deve ser "201"
 
   Scenario: Fracasso no cadastro por campos obrigatórios não preenchidos
     Given o usuário deseja se cadastrar
@@ -35,7 +35,7 @@ Feature: Serviço de Cadastro de Usuários
     And ele deixa o campo "Confirmar Senha" com ""
     And ele envia uma requisição POST para "/api/cadastro"
     Then a resposta deve conter a mensagem "Confirmar Senha é obrigatório."
-    And o status code deve ser 400
+    And o status code deve ser "400"
 
   Scenario: Fracasso no cadastro por duplicação de ID única
     Given o usuário deseja se cadastrar
@@ -48,7 +48,7 @@ Feature: Serviço de Cadastro de Usuários
     And ele informa a confirmação da senha "Password456"
     And ele envia uma requisição POST para "/api/cadastro"
     Then a resposta deve conter a mensagem "Erro: email/cpf já está registrado."
-    And o status code deve ser 409
+    And o status code deve ser "409"
 
   Scenario: Fracasso no cadastro por senhas que não coincidem
     Given o usuário deseja se cadastrar
@@ -60,7 +60,7 @@ Feature: Serviço de Cadastro de Usuários
     And ele informa a confirmação da senha "DifferentPassword123"
     And ele envia uma requisição POST para "/api/cadastro"
     Then a resposta deve conter a mensagem "As senhas não coincidem."
-    And o status code deve ser 400
+    And o status code deve ser "400"
 
     Scenario: Fracasso no cadastro por formato inválido de email
     Given o usuário deseja se cadastrar
@@ -72,7 +72,7 @@ Feature: Serviço de Cadastro de Usuários
     And ele informa a confirmação da senha "SenhaForte123"
     And ele envia uma requisição POST para "/api/cadastro"
     Then a resposta deve conter a mensagem "Formato de email inválido. Use um email válido, como exemplo@dominio.com."
-    And o status code deve ser 400
+    And o status code deve ser "400"
 
   Scenario: Fracasso no cadastro por formato inválido de CPF
     Given o usuário deseja se cadastrar
@@ -84,7 +84,7 @@ Feature: Serviço de Cadastro de Usuários
     And ele informa a confirmação da senha "SenhaSegura456"
     And ele envia uma requisição POST para "/api/cadastro"
     Then a resposta deve conter a mensagem "CPF inválido. Digite um CPF válido no formato XXX.XXX.XXX-XX."
-    And o status code deve ser 400
+    And o status code deve ser "400"
   
   Scenario: Fracasso no cadastro por siape já registrado
     Given o usuário deseja se cadastrar
@@ -97,4 +97,5 @@ Feature: Serviço de Cadastro de Usuários
     And ele informa a confirmação da senha "Senha123"
     And ele envia uma requisição POST para "/api/cadastro"
     Then a resposta deve conter a mensagem "Erro: siape já está registrado."
+    And o status code deve ser "409"
